Get Menu categories from state
categories received from props required to filtered in order to return only the ones the user has access to. The attempt to improve performance by updating the state in componentDidUpdate worked well, but the if a user navigated away from the Map page and navigated back, the Menu was blank. The workaround was to perform the filtering in the render method. This changes remove the work around and make use of the method getDerivedStateFromProps